move claude tests
