+++ pkg/workflow/git_commands_test.go
@@ -6,6 +6,7 @@ import (
 
 func TestApplyDefaultGitCommandsForSafeOutputs(t *testing.T) {
 	compiler := NewCompiler(false, "", "test")
+	engine := NewClaudeEngine()
 
 	tests := []struct {
 		name        string
@@ -97,7 +98,7 @@ func TestApplyDefaultGitCommandsForSafeOutputs(t *testing.T) {
 
 			// Apply both default tool functions in sequence
 			tools = compiler.applyDefaultGitHubMCPTools(tools)
-			result := compiler.applyDefaultClaudeTools(tools, tt.safeOutputs)
+			result := engine.applyDefaultClaudeTools(tools, tt.safeOutputs)
 
 			// Check if claude section exists and has bash tool
 			claudeSection, hasClaudeSection := result["claude"]
@@ -169,6 +170,7 @@ func TestApplyDefaultGitCommandsForSafeOutputs(t *testing.T) {
 
 func TestAdditionalClaudeToolsForSafeOutputs(t *testing.T) {
 	compiler := NewCompiler(false, "", "test")
+	engine := NewClaudeEngine()
 
 	tests := []struct {
 		name               string
@@ -235,7 +237,7 @@ func TestAdditionalClaudeToolsForSafeOutputs(t *testing.T) {
 
 			// Apply both default tool functions in sequence
 			tools = compiler.applyDefaultGitHubMCPTools(tools)
-			result := compiler.applyDefaultClaudeTools(tools, tt.safeOutputs)
+			result := engine.applyDefaultClaudeTools(tools, tt.safeOutputs)
 
 			// Check if claude section exists
 			claudeSection, hasClaudeSection := result["claude"]
